# api/swagger.yaml
...
swagger: "2.0"
info:
  description: "The Open Service Broker API defines an HTTP(S) interface between Platforms\
    \ and Service Brokers."
  version: "master - might contain changes that are not yet released"
  title: "Open Service Broker API"
  contact:
    name: "Open Service Broker API"
    url: "https://www.openservicebrokerapi.org/"
    email: "open-service-broker-api@googlegroups.com"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: "localhost:3000"
schemes:
- "http"
- "https"
consumes:
- "application/json"
produces:
- "application/json"
security:
- basicAuth: []
paths:
...
  /v2/service_instances/{instance_id}/service_bindings/{binding_id}:
    get:
      tags:
      - "ServiceBindings"
      summary: "gets a service binding"
      operationId: "serviceBinding.get"
      parameters:
      - name: "X-Broker-API-Version"
        in: "header"
        description: "version number of the Service Broker API that the Platform will\
          \ use"
        required: true
        type: "string"
      - name: "X-Broker-API-Originating-Identity"
        in: "header"
        description: "identity of the user that initiated the request from the Platform"
        required: false
        type: "string"
      - name: "X-Broker-API-Request-Identity"
        in: "header"
        description: "idenity of the request from the Platform"
        required: false
        type: "string"
      - name: "instance_id"
        in: "path"
        description: "instance id of instance to provision"
        required: true
        type: "string"
      - name: "binding_id"
        in: "path"
        description: "binding id of binding to create"
        required: true
        type: "string"
      - name: "service_id"
        in: "query"
        description: "id of the service associated with the instance"
        required: false
        type: "string"
      - name: "plan_id"
        in: "query"
        description: "id of the plan associated with the instance"
        required: false
        type: "string"
      responses:
        "200":
          description: "OK"
          schema:
            $ref: "#/definitions/ServiceBindingResource"
        "401":
          description: "Unauthorized"
          schema:
            $ref: "#/definitions/Error"
        "404":
          description: "Not Found"
          schema:
            $ref: "#/definitions/Error"
        "412":
          description: "Precondition Failed"
          schema:
            $ref: "#/definitions/Error"
        default:
          description: "Unexpected"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "ServiceBindings"
    put:
      tags:
      - "ServiceBindings"
      summary: "generation of a service binding"
      operationId: "serviceBinding.binding"
      parameters:
      - name: "X-Broker-API-Version"
        in: "header"
        description: "version number of the Service Broker API that the Platform will\
          \ use"
        required: true
        type: "string"
      - name: "X-Broker-API-Originating-Identity"
        in: "header"
        description: "identity of the user that initiated the request from the Platform"
        required: false
        type: "string"
      - name: "X-Broker-API-Request-Identity"
        in: "header"
        description: "idenity of the request from the Platform"
        required: false
        type: "string"
      - name: "instance_id"
        in: "path"
        description: "instance id of instance to provision"
        required: true
        type: "string"
      - name: "binding_id"
        in: "path"
        description: "binding id of binding to create"
        required: true
        type: "string"
      - in: "body"
        name: "body"
        description: "parameters for the requested service binding"
        required: true
        schema:
          $ref: "#/definitions/ServiceBindingRequest"
      - name: "accepts_incomplete"
        in: "query"
        description: "asynchronous operations supported"
        required: false
        type: "boolean"
      responses:
        "200":
          description: "OK"
          schema:
            $ref: "#/definitions/ServiceBindingResponse"
        "201":
          description: "Created"
          schema:
            $ref: "#/definitions/ServiceBindingResponse"
        "202":
          description: "Accepted"
          schema:
            $ref: "#/definitions/AsyncOperation"
        "400":
          description: "Bad Request"
          schema:
            $ref: "#/definitions/Error"
        "401":
          description: "Unauthorized"
          schema:
            $ref: "#/definitions/Error"
        "409":
          description: "Conflict"
          schema:
            $ref: "#/definitions/Error"
        "410":
          description: "Gone"
          schema:
            $ref: "#/definitions/Error"
        "412":
          description: "Precondition Failed"
          schema:
            $ref: "#/definitions/Error"
        "422":
          description: "Unprocessable Entity"
          schema:
            $ref: "#/definitions/Error"
        default:
          description: "Unexpected"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "ServiceBindings"
    delete:
      tags:
      - "ServiceBindings"
      summary: "deprovision of a service binding"
      operationId: "serviceBinding.unbinding"
      parameters:
      - name: "X-Broker-API-Version"
        in: "header"
        description: "version number of the Service Broker API that the Platform will\
          \ use"
        required: true
        type: "string"
      - name: "X-Broker-API-Originating-Identity"
        in: "header"
        description: "identity of the user that initiated the request from the Platform"
        required: false
        type: "string"
      - name: "X-Broker-API-Request-Identity"
        in: "header"
        description: "idenity of the request from the Platform"
        required: false
        type: "string"
      - name: "instance_id"
        in: "path"
        description: "instance id of instance to provision"
        required: true
        type: "string"
      - name: "binding_id"
        in: "path"
        description: "binding id of binding to create"
        required: true
        type: "string"
      - name: "service_id"
        in: "query"
        description: "id of the service associated with the instance being deleted"
        required: true
        type: "string"
      - name: "plan_id"
        in: "query"
        description: "id of the plan associated with the instance being deleted"
        required: true
        type: "string"
      - name: "accepts_incomplete"
        in: "query"
        description: "asynchronous operations supported"
        required: false
        type: "boolean"
      responses:
        "200":
          description: "OK"
          schema:
            type: "object"
        "202":
          description: "Accepted"
          schema:
            $ref: "#/definitions/AsyncOperation"
        "400":
          description: "Bad Request"
          schema:
            $ref: "#/definitions/Error"
        "401":
          description: "Unauthorized"
          schema:
            $ref: "#/definitions/Error"
        "410":
          description: "Gone"
          schema:
            $ref: "#/definitions/Error"
        "412":
          description: "Precondition Failed"
          schema:
            $ref: "#/definitions/Error"
        "422":
          description: "Unprocessable Entity"
          schema:
            $ref: "#/definitions/Error"
        default:
          description: "Unexpected"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "ServiceBindings"
securityDefinitions:
  basicAuth:
    type: "basic"
definitions:
...
  AsyncOperation:
    type: "object"
    properties:
      operation:
        type: "string"
...
  ServiceBindingResource:
    type: "object"
    properties:
      metadata:
        $ref: "#/definitions/ServiceBindingMetadata"
      credentials:
        type: "object"
        properties: {}
      syslog_drain_url:
        type: "string"
      route_service_url:
        type: "string"
      volume_mounts:
        type: "array"
        items:
          $ref: "#/definitions/ServiceBindingVolumeMount"
      endpoints:
        type: "array"
        items:
          $ref: "#/definitions/ServiceBindingEndpoint"
      parameters:
        type: "object"
        properties: {}
    example:
      metadata:
        expires_at: "expires_at"
        renew_before: "renew_before"
      route_service_url: "route_service_url"
      endpoints:
      - protocol: "tcp"
        host: "host"
        ports:
        - "ports"
        - "ports"
      - protocol: "tcp"
        host: "host"
        ports:
        - "ports"
        - "ports"
      credentials: "{}"
      syslog_drain_url: "syslog_drain_url"
      volume_mounts:
      - mode: "r"
        driver: "driver"
        device_type: "shared"
        device:
          volume_id: "volume_id"
          mount_config: "{}"
        container_dir: "container_dir"
      - mode: "r"
        driver: "driver"
        device_type: "shared"
        device:
          volume_id: "volume_id"
          mount_config: "{}"
        container_dir: "container_dir"
      parameters: "{}"
  ServiceBindingRequest:
    type: "object"
    required:
    - "plan_id"
    - "service_id"
    properties:
      context:
        $ref: "#/definitions/Context"
      service_id:
        type: "string"
      plan_id:
        type: "string"
      app_guid:
        type: "string"
        x-deprecated: true
      bind_resource:
        $ref: "#/definitions/ServiceBindingResourceObject"
      parameters:
        type: "object"
        properties: {}
      predecessor_binding_id:
        type: "string"
    example:
      app_guid: "app_guid"
      predecessor_binding_id: "predecessor_binding_id"
      service_id: "service_id"
      context: {}
      bind_resource:
        app_guid: "app_guid"
        route: "route"
      parameters: "{}"
      plan_id: "plan_id"
  ServiceBindingResourceObject:
    type: "object"
    properties:
      app_guid:
        type: "string"
      route:
        type: "string"
    example:
      app_guid: "app_guid"
      route: "route"
  ServiceBindingResponse:
    type: "object"
    properties:
      metadata:
        $ref: "#/definitions/ServiceBindingMetadata"
      credentials:
        type: "object"
        properties: {}
      syslog_drain_url:
        type: "string"
      route_service_url:
        type: "string"
      volume_mounts:
        type: "array"
        items:
          $ref: "#/definitions/ServiceBindingVolumeMount"
      endpoints:
        type: "array"
        items:
          $ref: "#/definitions/ServiceBindingEndpoint"
    example:
      metadata:
        expires_at: "expires_at"
        renew_before: "renew_before"
      route_service_url: "route_service_url"
      endpoints:
      - protocol: "tcp"
        host: "host"
        ports:
        - "ports"
        - "ports"
      - protocol: "tcp"
        host: "host"
        ports:
        - "ports"
        - "ports"
      credentials: "{}"
      syslog_drain_url: "syslog_drain_url"
      volume_mounts:
      - mode: "r"
        driver: "driver"
        device_type: "shared"
        device:
          volume_id: "volume_id"
          mount_config: "{}"
        container_dir: "container_dir"
      - mode: "r"
        driver: "driver"
        device_type: "shared"
        device:
          volume_id: "volume_id"
          mount_config: "{}"
        container_dir: "container_dir"
  ServiceBindingMetadata:
    type: "object"
    properties:
      expires_at:
        type: "string"
      renew_before:
        type: "string"
    example:
      expires_at: "expires_at"
      renew_before: "renew_before"
  ServiceBindingEndpoint:
    type: "object"
    required:
    - "host"
    - "ports"
    properties:
      host:
        type: "string"
      ports:
        type: "array"
        items:
          type: "string"
      protocol:
        type: "string"
        default: "tcp"
        enum:
        - "tcp"
        - "udp"
        - "all"
    example:
      protocol: "tcp"
      host: "host"
      ports:
      - "ports"
      - "ports"
  ServiceBindingVolumeMount:
    type: "object"
    required:
    - "container_dir"
    - "device"
    - "device_type"
    - "driver"
    - "mode"
    properties:
      driver:
        type: "string"
      container_dir:
        type: "string"
      mode:
        type: "string"
        enum:
        - "r"
        - "rw"
      device_type:
        type: "string"
        enum:
        - "shared"
      device:
        $ref: "#/definitions/ServiceBindingVolumeMountDevice"
    example:
      mode: "r"
      driver: "driver"
      device_type: "shared"
      device:
        volume_id: "volume_id"
        mount_config: "{}"
      container_dir: "container_dir"
  ServiceBindingVolumeMountDevice:
    type: "object"
    required:
    - "volume_id"
    properties:
      volume_id:
        type: "string"
      mount_config:
        type: "object"
        properties: {}
    example:
      volume_id: "volume_id"
      mount_config: "{}"
...
  Context:
    type: "object"
    description: "See [Context Conventions](https://github.com/openservicebrokerapi/servicebroker/blob/master/profile.md#context-object)\
      \ for more details."
...
  Error:
    type: "object"
    properties:
      error:
        type: "string"
      description:
        type: "string"
      instance_usable:
        type: "boolean"
      update_repeatable:
        type: "boolean"
...